tests and changelog
