+++ policyengine_canada/tests/gov/provinces/on/benefits/ocb/on_child_benefit_base.yaml
@@ -1,0 +1,20 @@
+- name: No children
+  period: 2023
+  input:
+    child_benefit_eligible_children: 0
+  output:
+    on_child_benefit_base: 0
+
+- name: One child
+  period: 2023
+  input:
+    child_benefit_eligible_children: 1
+  output:
+    on_child_benefit_base: 1_509
+
+- name: Two children
+  period: 2023
+  input:
+    child_benefit_eligible_children: 2
+  output:
+    on_child_benefit_base: 3_018
